enhance(bulk fastt): switch to config objects to configure the chart grid editor
